+++ .allowed-scripts.mjs
@@ -2,7 +2,7 @@ import { configureAllowedScripts } from '@ministryofjustice/hmpps-npm-script-all
 
 export default configureAllowedScripts({
   allowlist: {
-    'node_modules/cypress@15.17.0': 'ALLOW',
+    'node_modules/cypress@14.5.4': 'ALLOW',
     'node_modules/@parcel/watcher@2.5.6 ': 'ALLOW',
     // Provides native integration, supporting ability to write dtrace probes for bunyan
     'node_modules/dtrace-provider@0.8.8': 'ALLOW',
