+++ integration_tests/e2e/order/monitoring-conditions/alcoholMonitoring.cy.ts
@@ -1,9 +1,9 @@
 import { v4 as uuidv4 } from 'uuid'
-import { mockApiOrder } from '../../../mockApis/cemo'
 import ErrorPage from '../../../pages/error'
 import AlcoholMonitoringPage from '../../../pages/order/monitoring-conditions/alcohol-monitoring'
 import Page from '../../../pages/page'
 import InstallationLocationPage from '../../../pages/order/monitoring-conditions/installation-location'
+import mockApiOrder from '../../../utils/data/ApiOrder'
 
 const mockOrderId = uuidv4()
