+++ integration_tests/mockApis/cemo.ts
@@ -3,13 +3,13 @@ import { SuperAgentRequest } from 'superagent'
 import { v4 as uuidv4 } from 'uuid'
 import jsonDiff from 'json-diff'
 import { Client as PostgresqlClient } from 'pg'
-
 import logger from '../../logger'
 import { Order } from '../../server/models/Order'
 import { getMatchingRequests, stubFor } from './wiremock'
 import { DeviceWearer } from '../../server/models/DeviceWearer'
 import { DeviceWearerResponsibleAdult as ResponsibleAdult } from '../../server/models/DeviceWearerResponsibleAdult'
 import { OrderParameters } from '../../server/models/OrderParametersModel'
+import mockApiOrder from '../utils/data/ApiOrder'
 
 const ping = (httpStatus = 200) =>
   stubFor({
@@ -23,77 +23,6 @@ const ping = (httpStatus = 200) =>
       jsonBody: { status: httpStatus === 200 ? 'UP' : 'DOWN' },
     },
   })
-
-type ApiOrder = Omit<Order, 'deviceWearer'> & {
-  deviceWearer: Omit<Order['deviceWearer'], 'disabilities'> & {
-    disabilities: string
-  }
-}
-export const mockApiOrder = (status: Order['status'] = 'IN_PROGRESS', type: Order['type'] = 'REQUEST'): ApiOrder => ({
-  id: uuidv4(),
-  status,
-  type,
-  dataDictionaryVersion: 'DDV4',
-  deviceWearer: {
-    nomisId: null,
-    pncId: null,
-    deliusId: null,
-    prisonNumber: null,
-    homeOfficeReferenceNumber: null,
-    complianceAndEnforcementPersonReference: null,
-    courtCaseReferenceNumber: null,
-    firstName: null,
-    middleName: null,
-    lastName: null,
-    alias: null,
-    dateOfBirth: null,
-    adultAtTimeOfInstallation: null,
-    sex: null,
-    gender: null,
-    disabilities: '',
-    noFixedAbode: null,
-    interpreterRequired: null,
-  },
-  deviceWearerResponsibleAdult: null,
-  enforcementZoneConditions: [],
-  addresses: [],
-  contactDetails: null,
-  installationAndRisk: null,
-  probationDeliveryUnit: null,
-  interestedParties: null,
-  additionalDocuments: [],
-  monitoringConditions: {
-    orderType: null,
-    curfew: null,
-    exclusionZone: null,
-    trail: null,
-    mandatoryAttendance: null,
-    alcohol: null,
-    orderTypeDescription: null,
-    conditionType: null,
-    startDate: null,
-    endDate: null,
-    sentenceType: null,
-    issp: null,
-    hdc: null,
-    prarr: null,
-    pilot: null,
-    dapolMissedInError: null,
-    offenceType: null,
-    isValid: false,
-  },
-  monitoringConditionsTrail: null,
-  monitoringConditionsAlcohol: null,
-  mandatoryAttendanceConditions: [],
-  variationDetails: null,
-  isValid: false,
-  orderParameters: null,
-  versionId: uuidv4(),
-  dapoClauses: [],
-  offences: [],
-  mappa: null,
-  offenceAdditionalDetails: null,
-})
 
 type ListOrdersStubOptions = {
   httpStatus: number
@@ -802,7 +731,7 @@ const resetDB = async () => {
     await emptyNextTable(client)
   } catch (error) {
     // quite fail
-    logger.warn(`Error rest DB ${error} `)
+    logger.log(`Error rest DB ${error} `)
   }
   await client.end()
