+++ package.json
@@ -155,7 +155,7 @@
     "axe-core": "^4.11.0",
     "chokidar": "^3.6.0",
     "concurrently": "^9.2.1",
-    "cypress": "^15.17.0",
+    "cypress": "^14.5.4",
     "cypress-axe": "^1.7.0",
     "cypress-multi-reporters": "^2.0.5",
     "cypress-parallel": "^0.15.0",
@@ -177,7 +177,7 @@
     "micromatch": "^4.0.8",
     "mocha-junit-reporter": "^2.2.1",
     "nock": "^13.5.6",
-    "pg": "^8.16.3",
+    "pg": "^8.15.5",
     "prettier": "^3.6.2",
     "prettier-plugin-jinja-template": "^2.1.0",
     "secretlint": "^11.2.5",
