+++ server/services/orderService.test.ts
@@ -1,10 +1,10 @@
 import { v4 as uuidv4 } from 'uuid'
-import { mockApiOrder } from '../../integration_tests/mockApis/cemo'
 import { getMockOrder } from '../../test/mocks/mockOrder'
 import RestClient from '../data/restClient'
 import { SanitisedError } from '../sanitisedError'
 import OrderService from './orderService'
 import { VersionInformation } from '../models/VersionInformation'
+import mockApiOrder from '../../integration_tests/utils/data/ApiOrder'
 
 jest.mock('../data/restClient')
