+++ tsconfig.json
@@ -3,6 +3,7 @@
   "compilerOptions": {
     "target": "es2018",
     "module": "commonjs",
+    "moduleResolution": "node",
     "outDir": "./dist",
     "sourceMap": true,
     "skipLibCheck": true,
